Making todo list creator that uses the backend (#7)
* Going to make a hook for msal stuff

* Made new Hook

* Updated to use apiServices

- This allows me to not have to write the url in the usage place and instead have it in a centeralized area.

* Commmit to save new files

- Apparently an extension I'm using "VSpaceCode" doesn't actually stage
all changes when you use the magit git editor that has an option to stage all.
- Will need to look into

* Needing to rework axios-access token

- Right now the system I have only works for when loading a react
  compoent.
- Need to make sure it works when user clicks on a button

* Trying to do exhaustive deps

- Looks like this setup isn't giving me warnings about exhaustive deps

* Updating Todos Page

* Eslint not working right :/

* Removing eslint config since it didn't work

* Trying to restore hook

* Updated hook

* Fixed typing issue

* Calling endpoint from button

* Fixes small issue

* Perhaps fetch is good now

* Successfully sends and receives data

* Updating models correctly

* Needing to almost restart project for Azure AD B2C

- Apparently if I want a setup where any user can login with SSO on my
website I really need to use Azure AD B2C which is designed for that
kind of purpose. However, unfortunately the configuration settings
for each seem to be their own separate thing which means I'll
likely have to find a different guide/tutorial to help me setup the
configuration for the B2C version rather than the normal version.
- Note for clarity. Azure AD has been renamed to Microsoft Entra Id.
So any place where I mention Azure AD in the future (without the B2C)
I'm referring to the new "Microsoft Entra ID". It's just simply been
renamed.
- I think I will likely restart the entire project with new
project generations for the front-end and the backend and try to
find a matching tutorial for the B2C version.

* Deciding to archive repo instead of replace it

- I've decided I'm going to archive this repo and make a new git repo.